Replace grid formats with ListFormat
A `ListFormat[A]` is a function that reads the first line of a CSV file
and returns a `ColumnFormat[A]` to process the rest of the file.
This allows us to parse 2D grids as `Seqs` of case classes involving
fields of type `Map[String,???]`, which is a much more natural way
of processing many CSV formats than what we had before.